pyproject.toml
[project]
name = "facet-optimizer"
version = "0.1.0"
description = "Sample Braintrust facet executions, factor prompts into versioned YAML, build ground-truth datasets, and evaluate prompt changes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "autoevals>=0.0.129",
  "braintrust>=0.0.190",
  "openai>=1.100.0",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
